rename table-service for tbl-service
