+++ src/components/Utility/ContentTable/ContentTableHook.ts
@@ -2,7 +2,7 @@ import React from "react";
 import { Model, ModelFilter } from "@react3l/react3l/core";
 import { Dispatch, SetStateAction } from "react";
 import { importExportDataService } from "services/import-export-data-service";
-import tableService from "services/tbl-service";
+import tableService from "services/table-service";
 
 export function useContentTable<
   TContent extends Model, // mappingsField
